pyproject.toml
[tool.poetry]
name = "campus-admin-agent"
version = "0.1.0"
description = "AI-powered Campus Admin Agent with FastAPI and LangGraph"
authors = ["Saadi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
pydantic = {extras = ["email"], version = "^2.11.9"}
python-multipart = "^0.0.6"
sse-starlette = "^1.6.5"
python-dotenv = "^1.0.0"
openai = "^1.0.0"
motor = "^3.7.1"
beanie = "^2.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
